Update photo.gd
I added functionality to the photo object to access the database and confirmed it worked. I wrote notes of changes to make, like moving some of the functioanlity into a photo manager object and implementing many of the functions to work with tags. How I imagine it, Photo Manager will manage photos, and photos will manage tags.