is_favorited functionality working
I added a method to set_is_favorited to the opposite of what it currently is in the database and return the boolean value of what it was changed to. For example, it it's false and you call set_is_favorited(), it will set to and return true. I also wrote a method get_is_favorited().